Move IVFWriter to pkg
Allow this to be used/improved by others. SampleBuilder will be a
sibling

Resolves #115
